Move responsibilities of handling the keyboard show/hide from the FooterView to the QuizViewController since this is not the view responsibility. And since the FooterView was only doing this, it was remove from the project.
